25/04_161/Code/requirements.txt
torch
torchvision
transformers
opencv-python
numpy
pandas
Pillow
flask
